Melhorando o código usando double, para trabalhar com casas decimais

+++ adivinhacao.c
@@ -6,7 +6,7 @@ int main (){
     int numerosecreto = 42;
     int chute;
     int tentativas = 1;
-    int pontos = 1000;
+    double pontos = 1000;
 
     while (1){
         printf ("Tentativa %d\n", tentativas);
@@ -35,7 +35,7 @@ int main (){
 
         tentativas ++;
 
-        int pontosperdidos = (chute - numerosecreto) / 2;
+        double pontosperdidos = (chute - numerosecreto) / 2.0;
         pontos = pontos - pontosperdidos; 
     }    
     printf ("Fim de jogo!\n");
